requirements.txt
spacy>=3.7.0
python-dotenv>=1.0.1
google-genai>=1.0.0
pytest>=8.0.0
azure-identity>=1.17.1
langchain-openai>=0.3.0
openai>=1.45.0
requests>=2.32.0
